+++ src/trogdord/include/trogdord/io/input/listeners/redis.h
@@ -31,6 +31,12 @@ namespace input {
 			// Hostname we're connecting to
 			std::string hostname;
 
+			// Username used to authenticate (empty string if none)
+			std::string username;
+
+			// Password used to authenticate (empty string if none)
+			std::string password;
+
 			// The port redis is listening on
 			int port;